+++ src/main/java/com/netcracker/model/Car.java
@@ -18,7 +18,7 @@ public class Car {
     private String name;
 
     @Column(name = "volume")
-    private Integer volume;
+    private Double volume;
 
     @Column(name = "lifting_capacity")
     private Integer liftingCapacity;
@@ -73,11 +73,11 @@ public class Car {
         }
     }
 
-    public Integer getVolume() {
+    public Double getVolume() {
         return volume;
     }
 
-    public void setVolume(Integer volume) {
+    public void setVolume(Double volume) {
         if(volume != null) {
             this.volume = volume;
         }
